--- a/staging/src/k8s.io/pod-security-admission/policy/check_capabilities_baseline.go
+++ b/staging/src/k8s.io/pod-security-admission/policy/check_capabilities_baseline.go
@@ -22,6 +22,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/pod-security-admission/api"
 )
 
@@ -52,7 +53,7 @@ func CheckCapabilitiesBaseline() Check {
 		Versions: []VersionedCheck{
 			{
 				MinimumVersion: api.MajorMinorVersion(1, 0),
-				CheckPod:       capabilitiesBaseline_1_0,
+				CheckPod:       withOptions(capabilitiesBaselineV1Dot0),
 			},
 		},
 	}
@@ -76,34 +77,49 @@ var (
 	)
 )
 
-func capabilitiesBaseline_1_0(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec) CheckResult {
-	var badContainers []string
+func capabilitiesBaselineV1Dot0(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec, opts options) CheckResult {
+	badContainers := NewViolations(opts.withFieldErrors)
 	nonDefaultCapabilities := sets.NewString()
-	visitContainers(podSpec, func(container *corev1.Container) {
+	visitContainers(podSpec, opts, func(container *corev1.Container, path *field.Path) {
 		if container.SecurityContext != nil && container.SecurityContext.Capabilities != nil {
 			valid := true
-			for _, c := range container.SecurityContext.Capabilities.Add {
-				if !capabilities_allowed_1_0.Has(string(c)) {
-					valid = false
-					nonDefaultCapabilities.Insert(string(c))
+			if opts.withFieldErrors {
+				forbiddenValue := sets.NewString()
+				for _, c := range container.SecurityContext.Capabilities.Add {
+					if !capabilities_allowed_1_0.Has(string(c)) {
+						valid = false
+						nonDefaultCapabilities.Insert(string(c))
+						forbiddenValue.Insert(string(c))
+					}
 				}
-			}
-			if !valid {
-				badContainers = append(badContainers, container.Name)
+				if !valid {
+					badContainers.Add(container.Name, withBadValue(forbidden(path.Child("securityContext", "capabilities", "add")), forbiddenValue.List()))
+				}
+			} else {
+				for _, c := range container.SecurityContext.Capabilities.Add {
+					if !capabilities_allowed_1_0.Has(string(c)) {
+						valid = false
+						nonDefaultCapabilities.Insert(string(c))
+					}
+				}
+				if !valid {
+					badContainers.Add(container.Name)
+				}
 			}
 		}
 	})
 
-	if len(badContainers) > 0 {
+	if !badContainers.Empty() {
 		return CheckResult{
 			Allowed:         false,
 			ForbiddenReason: "non-default capabilities",
 			ForbiddenDetail: fmt.Sprintf(
 				"%s %s must not include %s in securityContext.capabilities.add",
-				pluralize("container", "containers", len(badContainers)),
-				joinQuote(badContainers),
+				pluralize("container", "containers", badContainers.Len()),
+				joinQuote(badContainers.Data()),
 				joinQuote(nonDefaultCapabilities.List()),
 			),
+			ErrList: badContainers.Errs(),
 		}
 	}
 	return CheckResult{Allowed: true}