--- a/staging/src/k8s.io/pod-security-admission/policy/check_capabilities_restricted.go
+++ b/staging/src/k8s.io/pod-security-admission/policy/check_capabilities_restricted.go
@@ -23,6 +23,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/pod-security-admission/api"
 )
 
@@ -63,29 +64,27 @@ func CheckCapabilitiesRestricted() Check {
 		Versions: []VersionedCheck{
 			{
 				MinimumVersion:   api.MajorMinorVersion(1, 22),
-				CheckPod:         capabilitiesRestricted_1_22,
+				CheckPod:         withOptions(capabilitiesRestrictedV1Dot22),
 				OverrideCheckIDs: []CheckID{checkCapabilitiesBaselineID},
 			},
 			// Starting 1.25, windows pods would be exempted from this check using pod.spec.os field when set to windows.
 			{
 				MinimumVersion:   api.MajorMinorVersion(1, 25),
-				CheckPod:         capabilitiesRestricted_1_25,
+				CheckPod:         withOptions(capabilitiesRestrictedV1Dot25),
 				OverrideCheckIDs: []CheckID{checkCapabilitiesBaselineID},
 			},
 		},
 	}
 }
 
-func capabilitiesRestricted_1_22(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec) CheckResult {
-	var (
-		containersMissingDropAll  []string
-		containersAddingForbidden []string
-		forbiddenCapabilities     = sets.NewString()
-	)
+func capabilitiesRestrictedV1Dot22(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec, opts options) CheckResult {
+	forbiddenCapabilities := sets.NewString()
+	containersMissingDropAll := NewViolations(opts.withFieldErrors)
+	containersAddingForbidden := NewViolations(opts.withFieldErrors)
 
-	visitContainers(podSpec, func(container *corev1.Container) {
+	visitContainers(podSpec, opts, func(container *corev1.Container, path *field.Path) {
 		if container.SecurityContext == nil || container.SecurityContext.Capabilities == nil {
-			containersMissingDropAll = append(containersMissingDropAll, container.Name)
+			containersMissingDropAll.Add(container.Name, required(path.Child("securityContext", "capabilities", "drop")))
 			return
 		}
 
@@ -97,32 +96,66 @@ func capabilitiesRestricted_1_22(podMetadata *metav1.ObjectMeta, podSpec *corev1
 			}
 		}
 		if !droppedAll {
-			containersMissingDropAll = append(containersMissingDropAll, container.Name)
+			if opts.withFieldErrors {
+				length := len(container.SecurityContext.Capabilities.Drop)
+				if length > 0 {
+					strSlice := make([]string, len(container.SecurityContext.Capabilities.Drop))
+					for i, v := range container.SecurityContext.Capabilities.Drop {
+						strSlice[i] = string(v)
+					}
+					forbiddenValues := sets.NewString(strSlice...)
+					containersMissingDropAll.Add(container.Name, withBadValue(forbidden(path.Child("securityContext", "capabilities", "drop")), forbiddenValues.List()))
+				} else if length == 0 {
+					containersMissingDropAll.Add(container.Name, required(path.Child("securityContext", "capabilities", "drop")))
+				}
+			} else {
+				containersMissingDropAll.Add(container.Name)
+			}
 		}
 
 		addedForbidden := false
-		for _, c := range container.SecurityContext.Capabilities.Add {
-			if c != capabilityNetBindService {
-				addedForbidden = true
-				forbiddenCapabilities.Insert(string(c))
+		if opts.withFieldErrors {
+			forbiddenValues := sets.NewString()
+			for _, c := range container.SecurityContext.Capabilities.Add {
+				if c != capabilityNetBindService {
+					addedForbidden = true
+					forbiddenCapabilities.Insert(string(c))
+					forbiddenValues.Insert(string(c))
+				}
+			}
+			if addedForbidden {
+				containersAddingForbidden.Add(container.Name, withBadValue(forbidden(path.Child("securityContext", "capabilities", "add")), forbiddenValues.List()))
+			}
+		} else {
+			for _, c := range container.SecurityContext.Capabilities.Add {
+				if c != capabilityNetBindService {
+					addedForbidden = true
+					forbiddenCapabilities.Insert(string(c))
+				}
+			}
+			if addedForbidden {
+				containersAddingForbidden.Add(container.Name)
 			}
 		}
-		if addedForbidden {
-			containersAddingForbidden = append(containersAddingForbidden, container.Name)
-		}
 	})
+
 	var forbiddenDetails []string
-	if len(containersMissingDropAll) > 0 {
+	var errList *field.ErrorList
+	if opts.withFieldErrors {
+		errs := append(*containersMissingDropAll.Errs(), *containersAddingForbidden.Errs()...)
+		errList = &errs
+	}
+	if !containersMissingDropAll.Empty() {
 		forbiddenDetails = append(forbiddenDetails, fmt.Sprintf(
 			`%s %s must set securityContext.capabilities.drop=["ALL"]`,
-			pluralize("container", "containers", len(containersMissingDropAll)),
-			joinQuote(containersMissingDropAll)))
+			pluralize("container", "containers", containersMissingDropAll.Len()),
+			joinQuote(containersMissingDropAll.Data())))
 	}
-	if len(containersAddingForbidden) > 0 {
+	if !containersAddingForbidden.Empty() {
 		forbiddenDetails = append(forbiddenDetails, fmt.Sprintf(
 			`%s %s must not include %s in securityContext.capabilities.add`,
-			pluralize("container", "containers", len(containersAddingForbidden)),
-			joinQuote(containersAddingForbidden),
+			pluralize("container", "containers", containersAddingForbidden.Len()),
+			joinQuote(containersAddingForbidden.Data()),
 			joinQuote(forbiddenCapabilities.List())))
 	}
 	if len(forbiddenDetails) > 0 {
@@ -130,16 +163,17 @@ func capabilitiesRestricted_1_22(podMetadata *metav1.ObjectMeta, podSpec *corev1
 			Allowed:         false,
 			ForbiddenReason: "unrestricted capabilities",
 			ForbiddenDetail: strings.Join(forbiddenDetails, "; "),
+			ErrList:         errList,
 		}
 	}
 	return CheckResult{Allowed: true}
 }
 
-func capabilitiesRestricted_1_25(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec) CheckResult {
+func capabilitiesRestrictedV1Dot25(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec, opts options) CheckResult {
 	// Pod API validation would have failed if podOS == Windows and if capabilities have been set.
 	// We can admit the Windows pod even if capabilities has not been set.
 	if podSpec.OS != nil && podSpec.OS.Name == corev1.Windows {
 		return CheckResult{Allowed: true}
 	}
-	return capabilitiesRestricted_1_22(podMetadata, podSpec)
+	return capabilitiesRestrictedV1Dot22(podMetadata, podSpec, opts)
 }