--- a/staging/src/k8s.io/pod-security-admission/policy/options.go
+++ b/staging/src/k8s.io/pod-security-admission/policy/options.go
@@ -1,0 +1,47 @@
+/*
+Copyright 2023 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package policy
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+type options struct {
+	withFieldErrors bool
+}
+
+type Option func(options) options
+
+func withOptions(f func(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec, opts options) CheckResult) CheckPodFn {
+	return func(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec, opts ...Option) CheckResult {
+		var opt options
+		for _, o := range opts {
+			if o != nil {
+				opt = o(opt)
+			}
+		}
+		return f(podMetadata, podSpec, opt)
+	}
+}
+
+func WithFieldErrors() Option {
+	return func(opt options) options {
+		opt.withFieldErrors = true
+		return opt
+	}
+}