@eddycharly 2023-04-15
There's something too complex here... and error prone.

Let's take the `sysctls` check for example:
```golang
func sysctls(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec, sysctls_allowed_set sets.String, opts options) CheckResult {
	var forbiddenSysctls []string
	var errList field.ErrorList

	if podSpec.SecurityContext != nil {
		for i, sysctl := range podSpec.SecurityContext.Sysctls {
			if !sysctls_allowed_set.Has(sysctl.Name) {
				forbiddenSysctls = append(forbiddenSysctls, sysctl.Name)
				opts.errListHandler(func() {
					err := withBadValue(field.Forbidden(sysctlsPath.Index(i).Child("name"), ""), []string{
						sysctl.Name,
					})
					errList = append(errList, err)
				})
			}
		}
	}

	if len(forbiddenSysctls) > 0 {
		return CheckResult{
			Allowed:         false,
			ForbiddenReason: "forbidden sysctls",
			ForbiddenDetail: strings.Join(forbiddenSysctls, ", "),
			ErrList:         errList,
		}
	}
	return CheckResult{Allowed: true}
}
```

This check is simple, yet it's going to be difficult to keep `forbiddenSysctls` and `errList` in sync. To me it looks like the field error and the forbidden sync should be stored together to reduce the risk adding one but not the other.

How about adding a bit of tooling to make this easier ?
We could go with something like this:
```golang

type violations[T any] struct {
	data  []T
	errs  field.ErrorList
	errFn func(*field.Path, T) *field.Error
}

func (v *violations[T]) Add(data T, pathFn func() *field.Path) {
	v.data = append(v.data, data)
	if v.errFn != nil {
		var path *field.Path
		if pathFn != nil {
			path = pathFn()
		}
		if path != nil {
			if err := v.errFn(path, data); err != nil {
				v.errs = append(v.errs, err)
			}
		}
	}
}

func (v *violations[T]) Empty() bool {
	return len(v.data) == 0 && len(v.errs) == 0
}

func (v *violations[T]) Data() []T {
	return v.data
}

func (v *violations[T]) Errs() field.ErrorList {
	return v.errs
}

type PathFn func() *field.Path

func withPath(path *field.Path) PathFn {
	if path == nil {
		return nil
	}
	return func() *field.Path {
		return path
	}
}

func (parent PathFn) index(i int) PathFn {
	if parent == nil {
		return nil
	}
	return func() *field.Path {
		parent := parent()
		if parent == nil {
			return nil
		}
		return parent.Index(i)
	}
}

func (parent PathFn) child(name string) PathFn {
	if parent == nil {
		return nil
	}
	return func() *field.Path {
		parent := parent()
		if parent == nil {
			return nil
		}
		return parent.Child(name)
	}
}
```

And the `sysctls` check would become:
```golang
func sysctls(podMetadata *metav1.ObjectMeta, podSpec *corev1.PodSpec, sysctls_allowed_set sets.String, opts options) CheckResult {
	forbiddenSysctls := violations[string]{
		errFn: func(path *field.Path, value string) *field.Error {
			return withBadValue(field.Forbidden(path, ""), value)
		},
	}

	pathFn := withPath(sysctlsPath)
	if podSpec.SecurityContext != nil {
		for i, sysctl := range podSpec.SecurityContext.Sysctls {
			if !sysctls_allowed_set.Has(sysctl.Name) {
				forbiddenSysctls.Add(sysctl.Name, pathFn.index(i).child("name"))
			}
		}
	}

	if !forbiddenSysctls.Empty() {
		return CheckResult{
			Allowed:         false,
			ForbiddenReason: "forbidden sysctls",
			ForbiddenDetail: strings.Join(forbiddenSysctls.Data(), ", "),
			ErrList:         forbiddenSysctls.Errs(),
		}
	}
	return CheckResult{Allowed: true}
}
```

With this, most of the logic is encapsulated in the `violations` struct, it becomes a lot harder to make mistakes and makes the code nicely reusable.
Of course, if `errFn` is `nil` no field errors will be reported, if we plumb this with the options logic enabling/disabling field errors becomes trivial.

cc @YTGhost @tallclair @liggitt WDYT ?