@liggitt 2023-05-17
queued up to take a look at the functional/structural changes, but I started by checking the impact on the existing benchmarks for pod / deployment / namespace validation when not using field paths:

```sh
git checkout master
go test ./plugin/pkg/admission/security/podsecurity -bench . -benchmem -run X -count=10 > old.txt

git checkout extend-psa
go test ./plugin/pkg/admission/security/podsecurity -bench . -benchmem -run X -count=10 > new.txt

go install golang.org/x/perf/cmd/benchstat@latest
benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: k8s.io/kubernetes/plugin/pkg/admission/security/podsecurity
                                                       │   old.txt    │               new.txt                │
                                                       │    sec/op    │    sec/op     vs base                │
VerifyPod/enforce-implicit_pod-10                        296.2n ±  2%   331.8n ± 16%  +12.00% (p=0.000 n=10)
VerifyPod/enforce-implicit_deployment-10                 131.3n ±  0%   128.6n ±  0%   -2.06% (p=0.001 n=10)
VerifyPod/enforce-privileged_pod-10                      314.3n ±  1%   311.2n ±  1%   -0.99% (p=0.000 n=10)
VerifyPod/enforce-privileged_deployment-10               153.4n ±  0%   149.9n ±  0%   -2.28% (p=0.000 n=10)
VerifyPod/enforce-baseline_pod-10                        2.910µ ± 18%   3.351µ ±  2%        ~ (p=0.101 n=10)
VerifyPod/enforce-baseline_deployment-10                 2.556µ ±  3%   2.569µ ±  1%        ~ (p=0.197 n=10)
VerifyPod/enforce-restricted_pod-10                      2.946µ ± 13%   3.597µ ±  1%  +22.12% (p=0.000 n=10)
VerifyPod/enforce-restricted_deployment-10               2.188µ ±  1%   2.780µ ±  1%  +27.09% (p=0.000 n=10)
VerifyPod/warn-baseline_pod-10                           2.630µ ±  1%   3.107µ ±  1%  +18.12% (p=0.000 n=10)
VerifyPod/warn-baseline_deployment-10                    2.122µ ±  1%   2.560µ ±  1%  +20.62% (p=0.000 n=10)
VerifyPod/warn-restricted_pod-10                         2.717µ ±  1%   3.382µ ±  1%  +24.45% (p=0.000 n=10)
VerifyPod/warn-restricted_deployment-10                  2.201µ ±  1%   2.804µ ±  1%  +27.40% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_pod-10             2.819µ ±  1%   3.305µ ±  1%  +17.26% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_deployment-10      2.141µ ±  1%   2.550µ ±  1%  +19.11% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_pod-10          3.705µ ±  1%   4.848µ ±  1%  +30.87% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_deployment-10   3.106µ ±  2%   4.158µ ±  1%  +33.89% (p=0.000 n=10)
VerifyNamespace-10                                       5.738m ±  7%   7.878m ±  3%  +37.29% (p=0.000 n=10)
geomean                                                  2.278µ         2.665µ        +17.02%

                                                       │   old.txt    │                new.txt                 │
                                                       │     B/op     │     B/op      vs base                  │
VerifyPod/enforce-implicit_pod-10                          192.0 ± 0%     192.0 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-implicit_deployment-10                   112.0 ± 0%     112.0 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-privileged_pod-10                        192.0 ± 0%     192.0 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-privileged_deployment-10                 112.0 ± 0%     112.0 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-baseline_pod-10                        4.174Ki ± 0%   5.097Ki ± 0%  +22.11% (p=0.000 n=10)
VerifyPod/enforce-baseline_deployment-10                 3.828Ki ± 0%   4.742Ki ± 0%  +23.88% (p=0.000 n=10)
VerifyPod/enforce-restricted_pod-10                      4.229Ki ± 0%   5.214Ki ± 0%  +23.30% (p=0.000 n=10)
VerifyPod/enforce-restricted_deployment-10               3.875Ki ± 0%   4.859Ki ± 0%  +25.40% (p=0.000 n=10)
VerifyPod/warn-baseline_pod-10                           4.072Ki ± 0%   4.995Ki ± 0%  +22.66% (p=0.000 n=10)
VerifyPod/warn-baseline_deployment-10                    3.828Ki ± 0%   4.742Ki ± 0%  +23.88% (p=0.000 n=10)
VerifyPod/warn-restricted_pod-10                         4.119Ki ± 0%   5.104Ki ± 0%  +23.92% (p=0.000 n=10)
VerifyPod/warn-restricted_deployment-10                  3.875Ki ± 0%   4.859Ki ± 0%  +25.40% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_pod-10             4.174Ki ± 0%   5.097Ki ± 0%  +22.11% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_deployment-10      3.828Ki ± 0%   4.742Ki ± 0%  +23.88% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_pod-10          5.573Ki ± 0%   7.481Ki ± 0%  +34.24% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_deployment-10   5.328Ki ± 0%   7.227Ki ± 0%  +35.63% (p=0.000 n=10)
VerifyNamespace-10                                       5.709Mi ± 0%   8.793Mi ± 0%  +54.03% (p=0.000 n=10)
geomean                                                  2.908Ki        3.501Ki       +20.39%
¹ all samples are equal

                                                       │   old.txt   │                 new.txt                 │
                                                       │  allocs/op  │  allocs/op    vs base                   │
VerifyPod/enforce-implicit_pod-10                         3.000 ± 0%     3.000 ± 0%         ~ (p=1.000 n=10) ¹
VerifyPod/enforce-implicit_deployment-10                  1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
VerifyPod/enforce-privileged_pod-10                       3.000 ± 0%     3.000 ± 0%         ~ (p=1.000 n=10) ¹
VerifyPod/enforce-privileged_deployment-10                1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
VerifyPod/enforce-baseline_pod-10                         29.00 ± 0%     44.00 ± 0%   +51.72% (p=0.000 n=10)
VerifyPod/enforce-baseline_deployment-10                  24.00 ± 0%     39.00 ± 0%   +62.50% (p=0.000 n=10)
VerifyPod/enforce-restricted_pod-10                       30.00 ± 0%     52.00 ± 0%   +73.33% (p=0.000 n=10)
VerifyPod/enforce-restricted_deployment-10                25.00 ± 0%     47.00 ± 0%   +88.00% (p=0.000 n=10)
VerifyPod/warn-baseline_pod-10                            28.00 ± 0%     43.00 ± 0%   +53.57% (p=0.000 n=10)
VerifyPod/warn-baseline_deployment-10                     24.00 ± 0%     39.00 ± 0%   +62.50% (p=0.000 n=10)
VerifyPod/warn-restricted_pod-10                          29.00 ± 0%     51.00 ± 0%   +75.86% (p=0.000 n=10)
VerifyPod/warn-restricted_deployment-10                   25.00 ± 0%     47.00 ± 0%   +88.00% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_pod-10              29.00 ± 0%     44.00 ± 0%   +51.72% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_deployment-10       24.00 ± 0%     39.00 ± 0%   +62.50% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_pod-10           39.00 ± 0%     76.00 ± 0%   +94.87% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_deployment-10    35.00 ± 0%     72.00 ± 0%  +105.71% (p=0.000 n=10)
VerifyNamespace-10                                       63.06k ± 0%   141.06k ± 0%  +123.69% (p=0.000 n=10)
geomean                                                   22.96          35.26        +53.54%
¹ all samples are equal
```

That shows a pretty significant memory increase (20-30%) and allocation count increase (50-90%) in this PR even when field paths aren't being used. That needs to be ~eliminated to be able to consider this.