@YTGhost 2023-05-23
@liggitt We have made some optimizations and the current benchstat results are shown below, is this an acceptable result? cc @realshuting @eddycharly 
```
goos: darwin
goarch: arm64
pkg: k8s.io/kubernetes/plugin/pkg/admission/security/podsecurity
                                                       │   old.txt    │               new.txt                │
                                                       │    sec/op    │    sec/op     vs base                │
VerifyPod/enforce-implicit_pod-10                        290.7n ±  0%   302.9n ± 13%   +4.21% (p=0.001 n=10)
VerifyPod/enforce-implicit_deployment-10                 127.9n ±  1%   128.5n ±  1%   +0.47% (p=0.001 n=10)
VerifyPod/enforce-privileged_pod-10                      318.8n ±  0%   313.8n ±  0%   -1.57% (p=0.000 n=10)
VerifyPod/enforce-privileged_deployment-10               163.2n ± 52%   152.1n ±  2%   -6.77% (p=0.001 n=10)
VerifyPod/enforce-baseline_pod-10                        2.718µ ±  0%   3.057µ ±  1%  +12.45% (p=0.000 n=10)
VerifyPod/enforce-baseline_deployment-10                 2.051µ ±  1%   2.332µ ±  1%  +13.70% (p=0.000 n=10)
VerifyPod/enforce-restricted_pod-10                      2.847µ ±  1%   3.292µ ±  1%  +15.63% (p=0.000 n=10)
VerifyPod/enforce-restricted_deployment-10               2.147µ ±  1%   2.524µ ±  1%  +17.56% (p=0.000 n=10)
VerifyPod/warn-baseline_pod-10                           2.563µ ±  1%   2.842µ ±  0%  +10.89% (p=0.000 n=10)
VerifyPod/warn-baseline_deployment-10                    2.063µ ±  0%   2.307µ ±  1%  +11.83% (p=0.000 n=10)
VerifyPod/warn-restricted_pod-10                         2.622µ ±  1%   3.042µ ±  1%  +16.02% (p=0.000 n=10)
VerifyPod/warn-restricted_deployment-10                  2.131µ ±  0%   2.528µ ±  1%  +18.60% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_pod-10             2.741µ ±  0%   3.046µ ±  1%  +11.13% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_deployment-10      2.065µ ±  1%   2.333µ ±  1%  +13.01% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_pod-10          3.544µ ±  1%   4.322µ ±  1%  +21.94% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_deployment-10   3.015µ ±  2%   3.655µ ±  2%  +21.23% (p=0.000 n=10)
VerifyNamespace-10                                       6.486m ±  8%   7.073m ±  0%   +9.05% (p=0.000 n=10)
geomean                                                  2.218µ         2.459µ        +10.86%

                                                       │   old.txt    │                new.txt                 │
                                                       │     B/op     │     B/op      vs base                  │
VerifyPod/enforce-implicit_pod-10                          192.0 ± 0%     192.0 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-implicit_deployment-10                   112.0 ± 0%     112.0 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-privileged_pod-10                        192.0 ± 0%     192.0 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-privileged_deployment-10                 112.0 ± 0%     112.0 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-baseline_pod-10                        4.174Ki ± 0%   4.597Ki ± 0%  +10.13% (p=0.000 n=10)
VerifyPod/enforce-baseline_deployment-10                 3.828Ki ± 0%   4.250Ki ± 0%  +11.02% (p=0.000 n=10)
VerifyPod/enforce-restricted_pod-10                      4.229Ki ± 0%   4.714Ki ± 0%  +11.48% (p=0.000 n=10)
VerifyPod/enforce-restricted_deployment-10               3.875Ki ± 0%   4.359Ki ± 0%  +12.50% (p=0.000 n=10)
VerifyPod/warn-baseline_pod-10                           4.072Ki ± 0%   4.495Ki ± 0%  +10.38% (p=0.000 n=10)
VerifyPod/warn-baseline_deployment-10                    3.828Ki ± 0%   4.250Ki ± 0%  +11.02% (p=0.000 n=10)
VerifyPod/warn-restricted_pod-10                         4.119Ki ± 0%   4.604Ki ± 0%  +11.78% (p=0.000 n=10)
VerifyPod/warn-restricted_deployment-10                  3.875Ki ± 0%   4.359Ki ± 0%  +12.50% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_pod-10             4.174Ki ± 0%   4.597Ki ± 0%  +10.13% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_deployment-10      3.828Ki ± 0%   4.250Ki ± 0%  +11.02% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_pod-10          5.573Ki ± 0%   6.480Ki ± 0%  +16.28% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_deployment-10   5.328Ki ± 0%   6.234Ki ± 0%  +17.01% (p=0.000 n=10)
VerifyNamespace-10                                       5.709Mi ± 0%   7.265Mi ± 0%  +27.27% (p=0.000 n=10)
geomean                                                  2.908Ki        3.197Ki        +9.93%
¹ all samples are equal

                                                       │   old.txt   │                new.txt                │
                                                       │  allocs/op  │  allocs/op   vs base                  │
VerifyPod/enforce-implicit_pod-10                         3.000 ± 0%    3.000 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-implicit_deployment-10                  1.000 ± 0%    1.000 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-privileged_pod-10                       3.000 ± 0%    3.000 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-privileged_deployment-10                1.000 ± 0%    1.000 ± 0%        ~ (p=1.000 n=10) ¹
VerifyPod/enforce-baseline_pod-10                         29.00 ± 0%    33.00 ± 0%  +13.79% (p=0.000 n=10)
VerifyPod/enforce-baseline_deployment-10                  24.00 ± 0%    28.00 ± 0%  +16.67% (p=0.000 n=10)
VerifyPod/enforce-restricted_pod-10                       30.00 ± 0%    38.00 ± 0%  +26.67% (p=0.000 n=10)
VerifyPod/enforce-restricted_deployment-10                25.00 ± 0%    33.00 ± 0%  +32.00% (p=0.000 n=10)
VerifyPod/warn-baseline_pod-10                            28.00 ± 0%    32.00 ± 0%  +14.29% (p=0.000 n=10)
VerifyPod/warn-baseline_deployment-10                     24.00 ± 0%    28.00 ± 0%  +16.67% (p=0.000 n=10)
VerifyPod/warn-restricted_pod-10                          29.00 ± 0%    37.00 ± 0%  +27.59% (p=0.000 n=10)
VerifyPod/warn-restricted_deployment-10                   25.00 ± 0%    33.00 ± 0%  +32.00% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_pod-10              29.00 ± 0%    33.00 ± 0%  +13.79% (p=0.000 n=10)
VerifyPod/enforce-warn-audit-baseline_deployment-10       24.00 ± 0%    28.00 ± 0%  +16.67% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_pod-10           39.00 ± 0%    51.00 ± 0%  +30.77% (p=0.000 n=10)
VerifyPod/warn-baseline-audit-restricted_deployment-10    35.00 ± 0%    47.00 ± 0%  +34.29% (p=0.000 n=10)
VerifyNamespace-10                                       63.06k ± 0%   96.06k ± 0%  +52.33% (p=0.000 n=10)
geomean                                                   22.96         27.19       +18.42%
¹ all samples are equal
```